fix: resolve type mismatch, exception handling, hardcoded credentials, cost estimator, file system todos, and documentation issues
- api/routes/research.py: estimated_time_minutes int → float (matches float returned by cost estimator)
- database/connection.py: catch ArgumentError from make_url() alongside ValueError
- services/deep_cost_estimator.py: align revision_cycles to 3 (consistent with compare_research_depths); derive deep_cost_per_task dynamically
- tools/file_system.py: propagate JSON write failure in write_todos
- services/auth_service.py: remove hardcoded test credentials; correct return type to Optional[TokenResponse]
- DEEP_RESEARCH_GUIDE.md: fix timing (8-12 → 7-12 min), token range (~22,620), token %, Profit→Net Loss label, include_controversies → include_contradictions
- PROJECT_STATUS.md: replace stale Phase 4 start note with staging validation task; update Questions section
- updates.md: update date, completion % (~64 → ~85%), Phase 4 row in decision matrix, agent node description
- research_context JSON: lower credibility on unverified source (0.9→0.5), add status fields

+++ api/routes/research.py
@@ -34,7 +34,7 @@ class CreateResearchResponse(BaseModel):
     task_id: str
     status: str
     estimated_cost_usd: float
-    estimated_time_minutes: int
+    estimated_time_minutes: float
 
 
 class ResearchStatusResponse(BaseModel):
